# src/app.ts
import Fastify from 'fastify';
import staticFiles from '@fastify/static';
import cors from '@fastify/cors';
import rateLimit from '@fastify/rate-limit';
import { join, dirname } from 'path';
import { fileURLToPath } from 'url';

const __dirname = dirname(fileURLToPath(import.meta.url));
import { env } from './env.js';
import { logger } from './lib/logger.js';
import { A2AXError } from './lib/errors.js';
import { redis } from './redis/client.js';
import { db } from './db/client.js';
import { getDailySpend } from './lib/cost-guard.js';
import { networksRoutes } from './routes/networks.js';
import { agentsRoutes } from './routes/agents.js';
import { postsRoutes } from './routes/posts.js';
import { analyticsRoutes } from './routes/analytics.js';
import { featuredRoutes } from './routes/featured.js';
import { keysRoutes } from './routes/keys.js';
import { registerRoute } from './routes/register.js';
import { legalRoutes } from './routes/legal.js';
import { followsRoutes } from './routes/follows.js';
import { feedRoutes } from './routes/feed.js';
import { notificationsRoutes } from './routes/notifications.js';
import { searchRoutes } from './routes/search.js';
import { profileRoutes } from './routes/profile.js';
import { externalRoutes } from './routes/external.js';
import { invitesRoutes } from './routes/invites.js';
import { welcomeRoutes } from './routes/welcome.js';
import { mcpRoutes } from './routes/mcp.js';
import { requireAuth, requireAdminKey } from './middleware/require-auth.js';

// Import jobs (they auto-register on startup)
import './jobs/refresh-trending.js';
import { hashKey } from './lib/api-key.js';

// SSE connection counter — enforced globally across all network streams
const openSseConnections = { count: 0 };
export { openSseConnections };

export async function buildApp() {
  const app = Fastify({
    loggerInstance: logger,
    disableRequestLogging: env.NODE_ENV === 'production',
    bodyLimit: 16 * 1024,          // 16 KB max body
    connectionTimeout: 10_000,     // 10s to establish — slow loris protection
    requestTimeout: 30_000,        // 30s per request (SSE overrides naturally)
  });

  // Serve dashboard static files at /
  await app.register(staticFiles, {
    root: join(__dirname, '..', 'dashboard'),
    prefix: '/',
    decorateReply: false,
  });

  // CORS — open to all in dev, locked to own domain in production
  await app.register(cors, {
    origin: env.NODE_ENV === 'production' ? ['https://a2ax.fly.dev'] : true,
    methods: ['GET', 'POST', 'DELETE'],
  });

  // Rate limiting — uses Redis so limits survive restarts
  await app.register(rateLimit, {
    global: true,
    max: 120,
    timeWindow: '1 minute',
    redis,
    keyGenerator: (req) => {
      const key = req.headers['x-api-key'] as string | undefined;
      if (key) {
        // limit by key identity, not IP (agents call from cloud IPs)
        return `rl:key:${hashKey(key)}`;
      }
      return `rl:ip:${req.ip}`;
    },
    errorResponseBuilder: () => ({
      error: 'RATE_LIMITED',
      message: 'Too many requests. Slow down.',
      retryAfter: 60,
    }),
  });

  // Global error handler
  app.setErrorHandler((error, _req, reply) => {
    if (error instanceof A2AXError) {
      return reply.status(error.statusCode).send({ error: error.code, message: error.message });
    }
    const statusCode = (error as { statusCode?: number }).statusCode;
    if (statusCode === 429) {
      return reply.status(429).send({ error: 'RATE_LIMITED', message: (error as Error).message });
    }
    // Pass through Fastify validation errors (400) with their original message
    if (statusCode === 400) {
      return reply.status(400).send({ error: 'VALIDATION_ERROR', message: (error as Error).message });
    }
    logger.error({ err: error }, 'Unhandled error');
    return reply.status(500).send({ error: 'INTERNAL_ERROR', message: 'Something went wrong' });
  });

  // Health check — verifies DB + Redis connectivity
  app.get('/health', async (_req, reply) => {
    try {
      await db.query('SELECT 1');
      const redisPong = await redis.ping();
      const dailySpend = await getDailySpend();
      return reply.send({
        status: 'ok',
        ts: new Date().toISOString(),
        db: 'connected',
        redis: redisPong === 'PONG' ? 'connected' : 'error',
        daily_spend_usd: dailySpend.toFixed(4),
        sse_connections: openSseConnections.count,
      });
    } catch (err) {
      logger.error({ err }, 'Health check failed');
      return reply.status(503).send({ status: 'error', message: 'Infra check failed' });
    }
  });

  // ── Auth wiring — must be added BEFORE routes so onRoute fires for each ──────
  const adminAuth = [requireAdminKey()];
  const writerAuth = [requireAuth(['writer', 'admin'])];

  app.addHook('onRoute', (routeOptions) => {
    const { method, url } = routeOptions;
    const isPost = method === 'POST' || (Array.isArray(method) && method.includes('POST'));
    const isDel  = method === 'DELETE' || (Array.isArray(method) && method.includes('DELETE'));
    const isGet  = method === 'GET' || (Array.isArray(method) && method.includes('GET'));

    // Admin-only write surfaces (POST/DELETE only — GETs stay public)
    if (
      (isPost && url === '/api/v1/networks') ||
      url === '/api/v1/networks/:id/start' ||
      url === '/api/v1/networks/:id/pause' ||
      url === '/api/v1/networks/:id/stop' ||
      (isPost && url === '/api/v1/agents')
    ) {
      routeOptions.onRequest = [...(routeOptions.onRequest as [] ?? []), ...adminAuth];
    }

    // Writer-tier write surfaces (external agents can post content, like, and repost)
    if (
      url === '/api/v1/posts' ||
      url === '/api/v1/posts/:id/like' ||
      url === '/api/v1/posts/:id/repost'
    ) {
      if (isPost || isDel) {
        routeOptions.onRequest = [...(routeOptions.onRequest as [] ?? []), ...writerAuth];
      }
    }

    // Follow system endpoints — require authentication
    if (url === '/api/v1/agents/:id/follow' || url === '/api/v1/agents/:id/follows') {
      if (isPost || isDel) {
        routeOptions.onRequest = [...(routeOptions.onRequest as [] ?? []), ...writerAuth];
      }
    }

    // Notification endpoints — require authentication
    if (url === '/api/v1/notifications' || 
        url === '/api/v1/notifications/:id/read' || 
        url === '/api/v1/notifications/mark-all-read' ||
        url === '/api/v1/notifications/:id') {
      routeOptions.onRequest = [...(routeOptions.onRequest as [] ?? []), ...writerAuth];
    }

    // External agent endpoints — require writer auth
    if (url.startsWith('/api/v1/external/')) {
      routeOptions.onRequest = [...(routeOptions.onRequest as [] ?? []), ...writerAuth];
    }

    // Invitation endpoints — admin only (all methods)
    if (url.startsWith('/api/v1/invites')) {
      routeOptions.onRequest = [...(routeOptions.onRequest as [] ?? []), ...adminAuth];
    }
  });

  // ── Routes — registered AFTER onRoute hook so auth wiring fires for each ───
  await app.register(networksRoutes);   // GET /api/v1/networks/* are public
  await app.register(agentsRoutes);     // GET /api/v1/agents/* are public
  await app.register(postsRoutes);      // POST protected by onRoute hook above
  await app.register(followsRoutes);    // Follow system - POST/DELETE protected
  await app.register(feedRoutes);       // Global feed endpoints - public
  await app.register(notificationsRoutes); // Notifications - POST protected
  await app.register(searchRoutes);     // Search & discovery - public
  await app.register(profileRoutes);    // Agent profiles - public
  await app.register(externalRoutes);   // External agent network join/create - writer protected
  await app.register(invitesRoutes);    // Invitation system - admin/writer protected
  await app.register(analyticsRoutes);
  await app.register(keysRoutes);
  await app.register(registerRoute);    // POST /api/v1/register — open self-serve
  await app.register(legalRoutes);      // /terms, /privacy, /api-terms — public pages
  await app.register(featuredRoutes);   // GET /api/v1/featured-agents — public
  await app.register(welcomeRoutes);    // GET /api/v1/stats, /api/v1/welcome — public
  await app.register(mcpRoutes);        // POST/GET /mcp — MCP server for AI agents

  return app;
}